fix compile error

+++ src/ops/op_common/template/aicpu/kernel_launch.cc
@@ -421,8 +421,7 @@ extern "C" unsigned int HcclLaunchAicpuKernel(OpParam *param)
         HCCL_INFO("[HcclLaunchAicpuKernel] opUnfoldIdx[%llu] commName[%s] opType[%u] inputPtr[0x%016llx] inputSize[%llu] "
             "outputPtr[0x%016llx] outputSize[%llu] opMode[%u] algName[%s] isZeroCopy[%d] opExpanMode[%u] enableCache[%d]",
             opUnfoldIdx++, param->commName, static_cast<uint32_t>(param->opType),
-            static_cast<uint64_t>(param->inputPtr), param->inputSize,
-            static_cast<uint64_t>(param->outputPtr), param->outputSize,
+            param->inputPtr, param->inputSize, param->outputPtr, param->outputSize,
             static_cast<uint32_t>(param->opMode), param->algName, param->isZeroCopy,
             static_cast<uint32_t>(param->commOpExpansionMode), enableCache);
